Improved swipetool with labeling support

+++ app/src/main/java/eu/darken/bb/common/UnitConverter.kt
@@ -8,7 +8,7 @@ object UnitConverter {
         return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, dp, context.resources.displayMetrics).toInt()
     }
 
-    fun spToPx(context: Context, sp: Float): Int {
-        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, sp, context.resources.displayMetrics).toInt()
+    fun spToPx(context: Context, sp: Float): Float {
+        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_SP, sp, context.resources.displayMetrics)
     }
 }
